refactor: ヘッダーを削除し言語切り替えをPassportFormに移動
- App.tsxの不要なヘッダー部分を削除
- 言語切り替えボタンをPassportFormのボタン群に移動
- より統合されたUI設計に変更
